Moved canvas list creation to new file
